pyproject.toml
[tool.poetry]
name = "angra"
version = "0.1.0"
description = ""
authors = ["mtsz-thiago <mtsz.thiago@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
dvc = "^2.5.4"
boto3 = "^1.18.1"
biopython = "^1.79"
ipykernel = "^6.0.2"
pandas = "^1.3.0"
awscliv2 = "^1.5.2"

[tool.poetry.dev-dependencies]
python = "^3.8"
dvc = "^2.5.4"
boto3 = "^1.18.1"
biopython = "^1.79"
ipykernel = "^6.0.2"
pandas = "^1.3.0"
awscliv2 = "^1.5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
